Add identifier to taxonomy icons
